+++ src/test/java/de/muspellheim/commons/sql/BrokenEntity.java
@@ -1,0 +1,21 @@
+/*
+ * Muspellheim Commons
+ * Copyright (c) 2019 Falko Schumann
+ */
+
+package de.muspellheim.commons.sql;
+
+import javax.swing.*;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@SuppressWarnings("checkstyle:VisibilityModifier")
+public class BrokenEntity {
+
+    String text;
+    JPanel panel;
+
+}